# docs/swagger.yaml
swagger: '2.0'
info:
  title: Gryphon API
  description: Store and load case management fragments using BPMN and a JSON API.
  version: 1.0.0
schemes:
  - http
basePath: /api
produces:
  - application/json
paths:
  /scenario/{id}:
    get:
      summary: Scenario
      description: |
        The scenario endpoint delivers information about a certain scenario, including
        revision, termination condition, all data objects and all fragments.
      parameters:
        - name: id
          in: path
          description: The scenario id.
          required: true
          type: number
          format: integer
        - name: populate
          in: query
          description: If true, the scenario will be populated with all child-objects.
          required: false
          type: boolean
      tags:
        - Scenario
      responses:
        '200':
          description: An scenario
          schema:
            $ref: '#/definitions/Scenario'
        '404':
          description: Scenario not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Updates a given scenario
      description: |
        If used, the repository will update the scenario that's using this idea. The repository doesn't care if the ID in the scenario object and the ID given via path match. The revision-attribute will be updated if there's an difference between the old and the new version.
      parameters:
        - name: id
          in: path
          description: The scenario id.
          required: true
          type: number
          format: integer
        - name: scenario
          in: body
          description: The scenario object.
          required: true
          schema:
            $ref: '#/definitions/Scenario'
      tags:
        - Scenario
      responses:
        '200':
          description: The updated scenario, possibly with updated revision.
          schema:
            $ref: '#/definitions/Scenario'
        '404':
          description: Scenario not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Deletes the given scenario
      description: |
        This endpoints deletes the given scenario. The associated domainmodel and
        fragments stay in the database.
      parameters:
        - name: id
          in: path
          description: The scenario id.
          required: true
          type: number
          format: integer
      tags:
        - Scenario
      responses:
        '200':
          description: The deleted scenario.
          schema:
            $ref: '#/definitions/Scenario'
        '404':
          description: Scenario not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /scenario:
    get:
      summary: Returns a list of all available scenarios.
      description: |
        Returns a list of all scenarios. This list might be filtered by searching
        using a name.
      parameters:
        - name: query
          in: query
          description: Might be used to filter the scenarios.
          required: false
          type: string
          format: string
      tags:
      - Scenario
      responses:
        '200':
          description: The list of scenarios
          schema:
            $ref: '#/definitions/ScenarioList'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Creates a new scenario
      description: |
        This endpoint stores a new scenario on the server. The DataObjects and
        Fragments will be added later using the associated endpoints. Any DomainModels and
        Fragments given in the objects server by the provided scenario object will be
        validated and added to the new scenario anyway.
      parameters:
        - name: scenario
          in: body
          description: The scenario object.
          required: true
          schema:
            $ref: '#/definitions/Scenario'
      tags:
        - Scenario
      responses:
        '200':
          description: The updated scenario, maybe with updated revision.
          schema:
            $ref: '#/definitions/Scenario'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /scenario/associatefragment:
    post:
      summary: Associated a fragment with a scenario.
      description: |
        This allows it to associate (add) a fragment with a scenario. To add a
        fragment to a scenario to steps are required. First upload the fragment
        using /fragment POST and then associate it with the scenario using this
        endpoint.
      parameters:
        - name: fragmentid
          in: query
          description: The fragment to associate.
          required: true
          type: number
          format: integer
        - name: scenarioid
          in: query
          description: The scenarios id.
          type: number
          format: integer
          required: true
      tags:
        - Fragment
        - Scenario
      responses:
        '200':
          description: The updated scenario if found.
          schema:
            $ref: '#/definitions/Scenario'
        '404':
          description: Fragment or scenario not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /scenario/associatedomainmodel:
    post:
      summary: Associates a domain model with a scenario.
      description: |
        This allows it to associate (add) a domain model with a scenario. To add
        a domain model to a scenario to steps are required. First upload the
        domain model using /domainmodel POST and then associate it with the
        scenario using this endpoint.
      parameters:
        - name: domainmodelid
          in: query
          description: The domainmodel to associate.
          type: number
          format: integer
          required: true
        - name: scenarioid
          in: query
          description: The scenarios id.
          type: number
          format: integer
          required: true
      tags:
        - DomainModel
        - Scenario
      responses:
        '200':
          description: The updated scenario if found.
          schema:
            $ref: '#/definitions/Scenario'
        '404':
          description: Domainmodel or scenario not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /scenario/{id}/validate:
    get:
      summary: Validates a scenario.
      description: |
        This endpoint validates a scenario. It validates all fragments according to the same rules that are used for the
        fragment-validation and validates all fragments.
      parameters:
        - name: id
          in: path
          description: The scenarios id.
          required: true
          type: number
          format: integer
      tags:
        - Scenario
        - Fragment
        - DomainModel
      responses:
        '200':
          description: The validation report.
          schema:
            $ref: '#/definitions/MessageList'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /scenario/{id}/export:
    get:
      summary: Exports a scenario.
      description: |
        This endpoint exports a scenario to the url given in the body object. Further additions to the body object might be possible.
        This endpoint exports the whole scenario (including fragments and domain model) in JSON format.
        The target server should respond with a 200 if everything went fine. The content of the response doesn't matter.
        The target server should expect a POST-call according to RESTful APIs.
      parameters:
        - name: id
          in: path
          description: The scenarios id.
          required: true
          type: number
          format: int32
        - name: exportID
          in: body
          description: The model object.
          required: true
          schema:
            $ref: '#/definitions/ExportRequest'
      tags:
        - Scenario
      responses:
        '200':
          description: The data that was returned by the target-server.
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Scenario doesn't exist
        '400':
          description: Target server is unreachable or doesn't respond as expected.
...
definitions:
  Scenario:
    type: object
    properties:
      name:
        type: string
        description: 'The scenarios name'
      _id:
        type: number
        format: integer
        description: 'Unique identifier of any scenario.'
      terminationconditions:
        type: array
        items:
          $ref: '#/definitions/TerminationCondition'
        description: 'The termination conditions that are used.'
      startconditions:
        type: array
        items:
          $ref: '#/definitions/StartCondition'
        description: 'The start conditions that trigger the start of this scenario'
      domainmodel:
        type: number
        format: integer
        description: 'The ID of the associated domain model.'
      fragments:
        type: array
        items:
          type: number
          format: integer
        description: 'The IDs of associated fragments.'
      revision:
        type: number
        format: integer
        description: |
          The scenarios revision. Is increased automatically if the
          fragments, the data_objects or the termination condition is different
          from the last version
  StartCondition:
    type: object
    properties:
      query:
        type: string
        description: 'The query that triggers this startcondition'
      dataclassses:
        type: array
        items:
          $ref: '#/definitions/StartConditionDataclass'
        description: 'The mapping of event-values to dataobject-values'
  StartConditionDataclass:
    type: object
    properties:
      classname:
        type: string
        description: 'The name of the dataclass, that should be used to instantiate this object'
      state:
        type: string
        description: 'The state the newly created dataobject has.'
      mapping:
        type: array
        items:
          $ref: '#/definitions/StartConditionAttrMapping'
        description: 'The definition which values of the event should be put in the values of this dataobject.'
  StartConditionAttrMapping:
    type: object
    properties:
      path:
        type: string
        description: 'The JSON Path that specifies where to read this value.'
      attr:
        type: string
        description: 'The attribute were the value should get written.'
  TerminationCondition:
    type: string
    description: |
      The termination condition in the format "Dataclass[State], Dataclass[State2]"
...
  MessageList:
    type: array
    description: 'A list of messages.'
    items:
      $ref: '#/definitions/Message'
  Message:
    type: object
    properties:
      text:
        type: string
        description: 'The message for this error. This should be displayed to the user.'
      type:
        type: string
        description: 'The type of the error.'
  ScenarioList:
    type: array
    description: 'All scenarios found.'
    items:
      $ref: '#/definitions/Scenario'
...
  Error:
    type: object
  ExportRequest:
    type: object
    properties:
      exportID:
        type: number
        format: int32
        description: The url the post-call should use.